requirements.txt
PyOpenGL
pygame
numpy
